hdl.rec: deprecate in favor of `lib.data` and `lib.wiring`.
Tracking #879.
