+++ amaranth/hdl/__init__.py
@@ -1,10 +1,14 @@
+import warnings
+
 from .ast import Shape, unsigned, signed
 from .ast import Value, Const, C, Mux, Cat, Repl, Array, Signal, ClockSignal, ResetSignal
 from .dsl import Module
 from .cd import ClockDomain
 from .ir import Elaboratable, Fragment, Instance
 from .mem import Memory
-from .rec import Record
+with warnings.catch_warnings():
+    warnings.filterwarnings(action="ignore", category=DeprecationWarning)
+    from .rec import Record
 from .xfrm import DomainRenamer, ResetInserter, EnableInserter
