+++ tests/test_hdl_rec.py
@@ -1,7 +1,11 @@
+import warnings
+
 from enum import Enum
 
 from amaranth.hdl.ast import *
-from amaranth.hdl.rec import *
+with warnings.catch_warnings():
+    warnings.filterwarnings(action="ignore", category=DeprecationWarning)
+    from amaranth.hdl.rec import *
 
 from .utils import *
